backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.11.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.5
Django==6.0.3
django-cors-headers==4.9.0
django-filter==25.2
django-phonenumber-field==8.4.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.8.0
idna==3.11
multidict==6.7.1
phonenumbers==9.0.25
pillow==12.1.1
propcache==0.4.1
PyJWT==2.11.0
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.5
twilio==9.10.2
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
